+++ cmd/hidden_lake/service/pkg/app/init_anon_node.go
@@ -77,7 +77,7 @@ func (p *sApp) initAnonNode() error {
 			}),
 			lru.NewLRUCache(hls_settings.CNetworkQueueCapacity),
 		),
-		queue.NewMessageQueueProcessor(
+		queue.NewQBTaskProcessor(
 			queue.NewSettings(&queue.SSettings{
 				FNetworkMask:          hls_settings.CNetworkMask,
 				FWorkSizeBits:         cfgSettings.GetWorkSizeBits(),