+++ pkg/network/anonymity/queue/queue_test.go
@@ -60,7 +60,7 @@ func testSettings(t *testing.T, n int) {
 func TestQueueVoidDisabled(t *testing.T) {
 	t.Parallel()
 
-	queue := NewMessageQueueProcessor(
+	queue := NewQBTaskProcessor(
 		NewSettings(&SSettings{
 			FNetworkMask:      1,
 			FWorkSizeBits:     10,
@@ -98,7 +98,7 @@ func TestRunStopQueue(t *testing.T) {
 		}),
 		asymmetric.LoadRSAPrivKey(testutils.Tc1PrivKey1024),
 	)
-	queue := NewMessageQueueProcessor(
+	queue := NewQBTaskProcessor(
 		NewSettings(&SSettings{
 			FMainPoolCapacity: testutils.TCQueueCapacity,
 			FRandPoolCapacity: 1,
@@ -122,7 +122,7 @@ func TestRunStopQueue(t *testing.T) {
 
 	err := testutils.TryN(50, 10*time.Millisecond, func() error {
 		sett := queue.GetSettings()
-		sQueue := queue.(*sMessageQueueProcessor)
+		sQueue := queue.(*sQBTaskProcessor)
 		if len(sQueue.fRandPool.fQueue) == int(sett.GetRandPoolCapacity()) {
 			return nil
 		}
@@ -165,7 +165,7 @@ func TestRunStopQueue(t *testing.T) {
 func TestQueue(t *testing.T) {
 	t.Parallel()
 
-	queue := NewMessageQueueProcessor(
+	queue := NewQBTaskProcessor(
 		NewSettings(&SSettings{
 			FNetworkMask:      1,
 			FWorkSizeBits:     10,
@@ -200,7 +200,7 @@ func TestQueue(t *testing.T) {
 	}
 }
 
-func testQueue(queue IMessageQueueProcessor) error {
+func testQueue(queue IQBTaskProcessor) error {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer func() {
 		cancel()